Packages (#943)
* Split off tournament clock to own module

* Get rid of model import in sound.ts

* Get rid of rollup warning  has been rewritten to

* ChessgroundController and GameController

* Fix move time chart

* Get rid of updateMaterial import in boardSettings

* Split off materialView

* Split off result functions

* Remove unnecessary lines

* Move getDests to gameCtrl as setDests

* Update requirements

* Update requirements

* Get rid of 'model' class member

* No need to export VariantConfig interface

* Remove outdated comment

* Fix Uncaught TypeError: this.doSendMove is not a function

* Use flipped() instead of using flip

* Split off type declarations to separate files

* Moved materialView stuff into roundCtrl

* Get rid of materialFunction

* Remove dodel from ChessgroundController

* Move material stuffs

* Make use of the flipped method

* Move the flipped method down
