+++ client/invite.ts
@@ -3,10 +3,10 @@ import { h, VNode } from 'snabbdom';
 import { _ } from './i18n';
 import { VARIANTS } from './chess';
 import { patch } from './document';
-import { gameType } from './profile';
+import { gameType } from './result';
 import { copyTextToClipboard } from './clipboard';
 import { timeControlStr } from './view';
-import { PyChessModel } from './main';
+import { PyChessModel } from './types';
 
 export function inviteView(model: PyChessModel): VNode[] {
     const gameId = model["gameId"];
