+++ client/movetimeChart.ts
@@ -2,7 +2,8 @@ import Highcharts from "highcharts";
 
 import { _ } from './i18n';
 import { selectMove } from './movelist';
-import AnalysisController from './analysisCtrl';
+import { AnalysisController } from './analysisCtrl';
+import { Step } from "./messages";
 
 export interface MovePoint {
   y: number;
@@ -35,7 +36,7 @@ export function movetimeChart(ctrl: AnalysisController) {
 
     const logC = Math.pow(Math.log(3), 2);
 
-    ctrl.steps.forEach((step, ply) => {
+    ctrl.steps.forEach((step: Step, ply: number) => {
         const turn = (ply + 1) >> 1;
         const color = ply & 1;
         const colorName = color ? 'white' : 'black';
@@ -50,8 +51,8 @@ export function movetimeChart(ctrl: AnalysisController) {
         if (ply <= 2) {step.clocks!.movetime = 0;
         } else {
             step.clocks!.movetime = (ply % 2 === 1) ?
-                (ctrl.steps[ply-1].clocks?.white! - ctrl.steps[ply].clocks?.white!) :
-                (ctrl.steps[ply-1].clocks?.black! - ctrl.steps[ply].clocks?.black!);
+                (ctrl.steps[ply-2].clocks?.white! - (ctrl.steps[ply].clocks?.white! - ctrl.inc * 1000)) :
+                (ctrl.steps[ply-2].clocks?.black! - (ctrl.steps[ply].clocks?.black! - ctrl.inc * 1000));
         }
 
         const y = Math.pow(Math.log(0.005 * Math.min(step.clocks!.movetime, 12e4) + 3), 2) - logC;
@@ -116,7 +117,7 @@ export function movetimeChart(ctrl: AnalysisController) {
         },
     };
 
-    const chart = Highcharts.chart('chart-movetime', {
+    ctrl.movetimeChart = Highcharts.chart('chart-movetime', {
         chart: { type: 'column',
             alignTicks: false,
             spacing: [2, 0, 2, 0],
@@ -245,8 +246,6 @@ export function movetimeChart(ctrl: AnalysisController) {
             },
         ]
     });
-
-    ctrl.movetimeChart = chart;
 }
 
 const formatClock = (movetime: number) => {
