+++ package.json
@@ -15,11 +15,11 @@
   },
   "funding": "patreon.com/pychess",
   "dependencies": {
-    "chessgroundx": "^9.1.9",
+    "chessgroundx": "^9.2.0",
     "ffish-es6": "^0.6.8",
     "fairy-stockfish-nnue.wasm": "^1.0.1",
     "gettext.js": "^1.1.1",
-    "highcharts": "^10.0.0",
+    "highcharts": "^10.1.0",
     "howler": "^2.2.3",
     "html-to-image": "^1.9.0",
     "snabbdom": "^3.1.0",
