+++ rollup.config.dev.js
@@ -5,6 +5,8 @@ const postcss = require('rollup-plugin-postcss')
 
 export default {
     input: "client/main.ts",
+    context: 'null',
+    moduleContext: 'null',
     output: {
         name: "PychessVariants",
         file: "static/pychess-variants.js",
